CA-1557: Delete any previous VG_XenSoures before performing a new one, since the VG names conflict.  We warn the user about this as appropriate (clean install only).
CA-1558: As a result of above change, we will no longer have this volume
lying around

CA-938: As a result of above change, the installation on the other disks will
be deleted.

Some of these changes may be reverted when we move to volume group per disk,
since we will no longer have problems with name clashes.