CA-2721.  Improve module loading code in the installer:
- make sure extramodules specified on the kernel command-line are also loaded
  by the dom0 (the specific fix here is making sure these modules are listed
  in the module-order file, which gets used to write dom0:/etc/modules).

- add an 'interactive-modules' option to 'init' to allow the user to confirm
  the loading of each module (useful for debugging/support).

- add 'blacklist' support.

This changeset reduces the total number of places 'modprobe' gets called from
down to 2 (one is usb-discover, and the other is the 'modprobe' function in
'init', calls to which replace os.system("modprobe %s" % module)).

Tested with XenRT job 9712, and also am testing further with a job targetted
at -m ALL to increase test coverage of this patch.