CA-4030: fixes for B/Zpreview upgrade
Now reads the xend domain config file to find out what the kernel commandline
root=... should be -- it's the only way to be sure. (In particular, booting
RHEL normally involves looking up the root= in xenstore, where it is put
by the custom installer. Tricky to access the xenstore db when booted from CD)

Also, the cached kernel and initrd names on B/Zpreview are transmitted to
the XE host where prepare-guest can stuff in the symlinks (see the upgrade.py
module, to be committed shortly)