CA-7090: Improve consistency of text across the installer screens:
Change the welcome screen text from:

   This setup tool will install %s on your server.

   This install will overwrite data on any hard drives you select to use during
   the install process. Please make sure you have backed up any data on this
   system be fore proceeding with the product install.

    %s[1] = PRODUCT_BRAND

to:

   This setup tool will install %s on your server.  Installing %s will erase
   all data on the disks selected for use unless an upgrade option is chosen.

   Please make sure you have backed up any data you wish to preserve before
   proceeding with the installation.

    %s[1,2] = PRODUCT_BRAND

Change the root password screen text from:

   Please specify the root password for this installation.

   (This is the password used when connecting to the %s from the %s.)

    %s[1] = BRAND_SERVER, %s[2] = BRAND_CONSOLE

Replace referecne to 'localised area' in timezone screen to 'city or area',
which is more accuract anyway and avoids the s/z spelling issue.

Switch to using a lower-case 'i' on the "Preparing for [I/i]nstallation..."
screen to be consistent with the final "Completing installation..." screen.