Simplify the startup script for the installer:
 - get rid of separate MODE_* and OPERATION_* enums.

 - unify answerfile and non-answerfile code-paths; simplify some other code
   paths.

Overall, this means that when users start a CD with the legacy option, they
will get placed directly into the P2V tool, without having to go through the
menu first.

This is the second version of this patch that fixes a problem with spontaneous
reboot that present in the first version.