Improve P2V client error handling.
Specifically, deprecate mount_os_root in favour of a simpler util.mount, then
wrap a number of operations that use mounted filesystems around try/finally so
that the mounts can be cleaned up properly.  Also, insert a umount of the main
filesystem at the end of the P2V script that was previously missing.