CA-22141: reapply 1162:8866b61c5f6c
I previously reverted this uneccesarily because of an installation
error that I didn't have time to track down and which might have been
caused by this changeset.

It now turns out that the problem was caused by my PXE
install.img/vmlinuz not matching the repository version, and nothing
to do with the changeset.