CA-25216: Bugfix: OEM installer disabled auto-NIC bringup if network used for install
The problem was that when the network was used for the installation
 (NFS/FTP/HTTP), the installer tried to configure firstboot
to set the mgmt NIC to be the NIC specified for installation.  But it
was forgot to save the NIC config, causing the default 'MODE=none'
line to get written into /etc/firstboot.d/data/interface-<MAC>.conf.