If mgmt interface is needed by an iscsi root disk then its ifcfg file (which is written by the installer) should have ONBOOT=yes.
This is not needed normally because the interface is brought up by
firstboot.  However, if the mgmt interface is needed by an iscsi
root disk we cannot tell firstboot to initialise it, as this would
cause dom0 to lose access to its root filesystem.

By setting ONBOOT=yes we tell /etc/init.d/network to bring up the
interface, and *this* script knows how to bring it up without taking
it down.

Without this change, if you install to an iSCSI disk the resulting
dom0 will have an invalid /etc/resolv.conf and not be able to resolve
DNS names.