Fixes for "Install to iSCSI" (/opt/xensource/installer/init --enable-iscsi)
Explicitly ask user which interface XenServer should use to connect to iSCSI
target and how it should be configured.  Then use this information
when creating the initrd.

Prior to this change, the installer would ask which interface it should
use to connect to the target during installation, and how that
interface should be configured during installation, but not which
interface it should use when it boots into XenServer, or how that
interface should be configured.  Instead it just assumed that the
interface would be the same as during installation, and that the
configuration would be DHCP.