CA-38209: support for install-to-multipath root in host-installer
Support install-to-multipath in host-installer unless --no_mpath is on
command line.

installer will use devscan.py (taken from SM backend) to determine
what disks if any are SAN-like and will multipath these.  For each of
these it will display info for only the first slave.

If this is a fresh install then the host-installer will tell firstboot
to enable multipathed SRs iff the root disk is multipathed.

For answerfile installs any slave can be specified and install to
multipath will proceed the same as if the first was specified.