CA-38304: Use --without-multipath with mkinitrd if primary-disk not multipath
Otherwise, if there are any dm-multipath devices present mkinitrd gets
confused and adds code to the initrd that tries to multipath the root
disk too.

This fixes the issue where at install time there is a choice between a
multipath disk and a non-multipath disk and the user selects the latter.