CA-38329: disallow SAN disks in Local SR (except for primary-disk)
It is not sensible to allow additional SAN disks to form the Local SR
volume group.  Also, if we were to allow this then since these disks are
multipathed, the DM node will be specified in the firstboot data.
This will cause /etc/firstboot.d/10-prepare-storage to fail because
the DM node will not exist at that point.