CA-38429: Fix installation onto CCISS disks
The extra option --rootdev=<partition> confuses mkinitrd if
<partition> == /dev/cciss/c0d0p1 with the result that it forgets to
add the cciss.ko module to the initrd.

This option is really only necessary if installing to a multipath root
disk (since /etc/fstab only specifies a LABEL and both the dm node and
the multipath slaves match).  So only append this option if the root
disk is multipathed.