CA-38419: Do not start multipathd syslog thread as this may cause deadlocks.
It looks likely that the syslog thread in multipathd occaisionally
deadlocks with the main thread.  This causes
diskutil.py:wait_for_multipathd() to timeout which in turn causes
occaisional installs to fail.

Running multipathd with the -d option makes it send all debug to
stdout instead of starting a syslog thread.

For the time being use the -d option (and redirect stdout).  If this fixes
the host install failures then the above analysis is probably correct and we
should try to find the exact cause.