Support "install to iSCSI disk listed in iBFT" (if --use_ibft option present)
If --use_ibft is specified on the host-installer command line the host
installer will:

 * insmod the iBFT kernel module
 * look for an iBFT in /sys/firmware/ibft
 * attach to any disks in the iBFT using the n/w configuration in the iBFT
 * multipath the iSCSI disks (there may be two paths in the iBFT for
   the same LUN)
 * offer the iSCSI disk as an install target as if it were locally attached

Any NICs which are used by disks in the iBFT will be reserved and not
offered for reconfiguration or management by the host installer,
or specified in the firstboot data to be introduced as PIFs in xapi.
The user is made aware of this at install time.

In requireNetworking we now check to see if any interfaces are already
up and if so offer the option of reusing the existing network
configuration rather than reconfiguring.  This is useful in the
general case but is neccessary if the install repository is accessed
via the same NIC as the root iSCSI LUN.