CA-38211,CA-38523: Use multipathd -e option to prevent maps being created at SoD
Use -e to prevent multipathd creating maps implicitly at start of
day.  This saves us from having to then delete the maps for HBAs
that we don't support with multipath.

This *may* also fix the problem that we sometimes install using
multipath when we shouldn't.  At least it will if the cause is that
the "remove map" command to multipathd is failing.