Fix installs to large disks.
DOS partition tables can only address 4 GigaSectors.  Therefore we need to
truncate so that only sector addresses in this range are passed to
sfdisk.  If we do not do this sfdisk fails.

This means that disks larger than this will have unused space at the
end.  This should be fixed in a later release when we support GPT.