Fix race: we need to ensure udev creates dev nodes before they're needed
There is a race condition in the installer.  sfdisk may cause new
partitions to be created, however we rely on udev creating dev nodes
for these quickly enough that they are in place before mkfs.ext3 is
called.

Use udevtrigger/udevsettle to remove the race.  At the moment we are
giving udevsettle a timeout of just 1 second because this was found to
fix a repro and because we are being cautious a.t.m.