Added disabled code to support GUID Partition Tables
Support for GPT is planned for Tampa (the release after Boston) but
the code is ready so it is being committed now, but in a disabled
state.

To enable GPT set GPT_SUPPORT to True in constants.py.

This will enable GPT on all new installs except for those where a DOS
Partition such as the Dell Utility Partition needs to be preserved.
On upgrades the pre-existing partition table type will be preserved
because it is not possible to upgrade from DOS to GPT without
reformatting the first and last partitions.

GPT support enables use of disks which are over 2TB.