EA-1012: Use two different styles of version.py to brand the host installer.
In the XS version.py, PRODUCT_BRAND='XenServer' and PRODUCT_VERSION is the current product version.
In XCP, PRODUCT_BRAND="" and PRODUCT_VERSION="". In both cases, PLATFORM_NAME="XCP" and
PLATFORM_VERSION is the current XCP version number. This patch causes various strings to use
PRODUCT_BRAND if present, and PLATFORM_NAME if not. Similarly for the versions.