CA-58511: is_iscsi(dev) should only return True if installer created session
Some device drivers (e.g. QLE4062c iSCSI Host Bus Adapter) are
partially integrated with open-iscsi and the kernel iscsi subsystem.
These may automatically connect to boot LUNs on insmod rather than
requiring an explicit iscsiadm command even though the sessions are
exposed in sysfs and viewable by iscsiadm.

If this sort of device is chosen for the root disk, we should make
is_iscsi() return False.  We have not explicitly connected to this
device and so there may be no iscsi settings in /etc that can be
copied to dom0.

On the other hand, if we have created the session in the host
installer we will need to copy the settings across to dom0 so that we
can chkconfig the iscsi daemon which handles multipath failovers.