CA-58939: [multipath root] cope with paths to devices coming up at any time
bug: if a path comes up late (after multipathd is started) it does not
get enslaved.  Then if an answerfile specifies a symlink in
/dev/disk/by-id it will point to an unenslaved device and the
installation will use a single path rather than a DM multipath root

fix: when multipathing is enabled create a udev rule which enslaves
devices as they come up.  This rule is processed before the rule that
creates the symlink to avoid races.