CA-48440: [multipath root] Cope with lost udev events when starting multipathd
Very occaisionally LUNs attach *after* multipathd has started (and so
they are not automatically enslaved) but *before* it is able to handle
the "path add" command issued by the udev rule.  In these cases the
LUNs will be left unenslaved.

To fix this we simply wait for multipathd to complete its
initialisation and then issue a "reconfigure" command in the
multipathd CLI.  This causes all multipathable devices to get enslaved.