CA-111154: Rolling Pool Upgrade finds connected NIC from bond
During a rolling pool upgrade the prepare_host_upgrade.py plugin
determines which network interface to use by passing a MAC to the
installer. Bond MACs default to that of the first slave, so if that
slave had failed / was disconnected, the upgrade would fail.

This change queries each slave in the bond until it finds one with
a network carrier, and uses that MAC for the installer.