Do not disable entirely IPv6 module loading
Although not used this module is required by bridge module so disabling
it entirely cause bridge network to not work.
The problem was introduced trying to update modprobe.conf syntax
(commit 3486ccb50c2a).
The old ipv6 disable stop working a while ago so before that patch it
used to work correctly (as was doing nothing).