CA-110802: Add map_netdev options to installer command line on upgrade
If the mapping of ethernet device name to MAC address is unusual, the
host installer on an upgrade might use a different mapping which will
prevent the install from working.

In the upgrade plugin, add map_netdev=ethN:d:<MAC> options to the
installer command line (via the kernel command line) for each ethernet
device.  This ensures the upgraded install ends up with the same
mapping.