CP-11220: Restore the restore functionality
For GPT disks, backup the partition table on upgrade and restore it on
restore so that any partition table changes that were performed on upgrade
are undone. This avoids getting into a situation where the primary partition
is reduced in size such that the backup is too large to fit.  Re-add support
for installing extlinux so that older installations can have the bootloader
restored.

For MBR disks, no partition changes are needed on upgrade or restore.