CA-174103: Use the same partition naming as with an installed system
The naming scheme for device mapper partitions for an installed
(CentOS 7) system is:
1) If the device ends with a number, add "pX" where X is the
partition number.
2) If the device does not end with a number, add "X" where X is the
partition number.

The host installer was passing flags to kpartx which altered this
behavior, causing the default SR to not be created on firstboot because
the partitions could not be found. Update the host installer to match
the installed system behavior.