CA-175430: Generate multipath conf when using multipath root
Previously, the custom dracut multipath module ran sm-multipath to
generate the multipath.conf. With the removal of the module, run
sm-multipath manually in the host installer when using a multipath root
device.  Also copy the wwids file from the installer to the installed
system since this file must be present in the initrd for "multipath -c"
to run properly when called from the udev rules.