CA-178617: Only copy active iSCSI nodes to install root
When the target is discovered during installation, node files are
created for all its interfaces, even those not specified in the iBFT. If
these are copied to the target system, iscsid will login to them
automatically on boot which is not desired.  To prevent this, only copy
node files from logged-in sessions.