CA-179282: Sync multipath configuration with SM
Update the multipath config file to match the one used by SM.

This fixes the following problem:
The existing one used a comma to terminate some lines. multipathd
interpreted the configuration incorrectly meaning that it was not being
applied. Because the configuration may cause different kernel modules to
be used (e.g. dm-round-robin), these modules were not being inserted
into the initrd and the host would consequently fail to boot.