Rename fcoe-interface and handle dcb config correctly
Answerfile processing was adding fcoe interfaces to fcoe-interface dict,
whereas tui was adding it to fcoe-interfaces dict. fcoe-interface is now
renamed to fcoe-interfaces.
dcb config present in fcoe-interfaces dict is a boolean and not a string. Fixed
all instance of code, which assumed dcb config to be a string.
Removed unsed code.