CA-199252: Prevent sfdisk disk-in-use error
A sequence of invocations of writing a partition table with sfdisk may
race with udev causing a disk-in-use error. It races with udev which
inspects the disk after it has been opened for writing to regenerate
disk/by-label links, etc.  This is exacerbated because sfdisk issues a
BLKRRPART ioctl (re-read partition table) before and after it changes
the partition table.

To fix, issue a udev settle before calling sfdisk to change the
partition table. Furthermore, prevent sfdisk from issuing BLKRRPART
before it changes the partition table.