CA-197142: Don't use udev trigger when writing partition table
udev trigger is an expensive operation intended to be used to replay all
device events from the kernel. On some systems, this can cause a lot of
work and cause the subsequent settle to timeout if a relatively short
timeout is used.

When writing the partition table, the desire is to wait for the device
nodes to appear. Since the kernel+udev will automatically detect
partition table changes, simply wait a bit to ensure that the change has
been detected then settle to ensure that all the udev scripts have
finished processing.