CA-202093: Remove 45-multipath.rules from host installer.
Disks with paritions never gets added to multipath because of the following
reason. udev takes shared (LOCK_SH) flock on the device nodes (/dev/sd[a-z]*)
when it receives an uevent. This will race with multipathd, which tries to take
exclusive (LOCK_EX) flock on the same device node.

Historically we have had a dynamic udev rule in host-installer, to add paths to
multipath if the paths appear after multipathd is start. This rule looks like
this. ACTION=="add", RUN+="/bin/sh -c 'echo add path %k | /sbin/multipathd -k.
But this rule will try to add partitions /dev/sd[a-z][0-9] as well to multipath,
which will be ignored by multipathd. So when the partitions are getting added
to mp by this rule, udev has already taken a LOCK_SH on the partition.
A lock on the partition is also visible on the parent device node, which means
when a lock on /dev/sde1 is taken multipathd can't lock /dev/sde.  This cause an
uevent to add /dev/sde to be lost, because udev has taken a lock on /dev/sde1.

We don't need this rule anymore, as the multipathd would receive an uevent
notification when new disks are added into the system. multipathd retires 5
times to take a lock, if it cant take a lock.