CA-204956: Handle conversion from scsi-* links to ata-* links
If a disk device has moved from using scsi-* link to ata-* links, update
the XAPI database and firstboot data to contain the correct links.

If the device in the XAPI database is incorrect, then the upgrade
plugin's safe2upgrade check may give incorect results, ultimately
causing the partition layout to be upgraded even though there are VDIs
on the local disk.