CA-204956: Improve safe2upgrade device matching
When looking for a local SR on the root disk, follow symlinks to ensure
the plugin is not fooled by different links to the same device being
used. Secondly, instead of checking for the parent disk using a simple
prefix check, use a regexp to prevent something like /dev/sdaa1 being
considered a partition of /dev/sda.