CA-210928: Fix restoring ownership on upgrade
When upgrading from one build to another, the mapping between
owner/group and uid/gid might change, depending on what packages are
installed and the order in which they are installed. This may result in
the restored files having the wrong owner/group. Copy the actual
owner/group across rather than just relying on the uid/gid.