CA-210846: Handle diskutil.idFromPartition returning None
diskutil.idFromPartition may return None if there is no by-id link for
the disk. It causes an AttributeError if one of the next conditions
succeeds (primary_disk starts with '/dev/disk/by-id/scsi-').  Handle
this correctly by using the original path instead.

This is a regression caused by 3405ae361833 ("CA-204956: Handle
conversion from scsi-* links to ata-* links").

Although this was seen when doing an invalid upgrade (from normal root
disk to multipath root disk), this may happen in other situations if
there is no by-id link for the disk.