CA-216099: Don't disable sshd at build time
A stale rule in the host installer RPM postinstall script disabled
sshd.service. However, depending on the order of installation between
openssh-server and host-installer (at build time), the rule may or may
not actually have had an effect. sshd.service is meant to be always
enabled, with an ExecStartPre rule determining whether it is actually
started based on the presence of sshpassword in the kernel command-line.