CP-17850: Update install code to support new supp pack format
With the groundwork in place, update the main installer logic to work
with the new supp pack format.

For answerfiles, multiple source lines can be listed to list a main repo
plus additional supp packs.  For interactive installs, after the main
installation is complete, the installer will prompt to install
additional supp packs.

With this change, installing driver disk packages is temporarily broken.