CA-228622: URL-encode and decode username/passwords
When a username/password is part of an URL (either http or ftp), the
username/password must be URL-encoded.  This is also the form that Yum
expects for its baseurl parameter.  Encode and decode username/passwords
where needed so that they are stored in the URL encoded but displayed
and used in decoded form.