CA-229477: Wait longer for FCoE devices to appear
With some FCoE setups (seen with an Intel 82599ES 10GbE NIC), it can
take more than 10 seconds for the FCoE block devices to appear (e.g. 12
seconds).  Furthermore, the devices can initially appear as LUNs without
a block device.  Workaround this by waiting a constant 30 seconds.
While this is a bit ugly, the wait only happens in the unlikely case
that an FCoE boot device is used.