CA-233146: Fix upgrade when interfaces names don't match
When inspecting old product installations, don't assume that the
management interface exists and has the same name since this may not
be true.  This fixes an error trying to get the MAC for a nonexistent
device. The MAC is not needed anyway.

This can happen if an RPU is performed where the source is accessed over
an interface that isn't eth0. A rule will be set up by the upgrade
plugin to ensure the interface has the same name in the install
environment. This causes interface-rename to rename the remaining 'new'
interfaces to have names higher than any static/dynamic rules.