CA-247345: Preserve log partition during upgrade (take 2)
If the log partition already exists, just update its label rather than
recreating it so that the log file from the previous installation are
preserved.

If the log partition has not been unmounted cleanly when it is
relabelled, it will revert to the original label once it is mounted. To
prevent this, fsck the log partition before relabelling.  If any
unfixable errors occur or relabelling fails, just recreate the
filesystem instead, rather than failing the installation.

Unlike the previous attempt, don't preserve the log partition for fresh
installs.