CA-250335: Update results dict in place
Update the results dictionary in place rather than updating a copy of
it. This fixes an issue where the logs from a failed installation are
not written to disk because the results dictionary does not contain the
install disk details (because it is not updated in place). This is
especially important for RPUs where there is often no console or serial
access.