CA-255453: Add /var/lib/xcp/blobs to the restore list
There is a compat symlink from /var/xapi to /var/lib/xcp but this
doesn't work with the restore code. It is because the symlink is
absolute but the filesystem is mounted in a subdirectory, so the symlink
appears broken during restore. The result is that /var/lib/xcp/blobs is
not restored when upgrading from a newer release which has /var/xapi as
a symlink.  Fix this by adding /var/lib/xcp/blobs to the restore list.