CA-255453: Speed up restoring files
6555af21a311 ("CA-210928: Fix restoring ownership on upgrade") changed
the restore process to restore ownership by name rather than id, as the
id numbers can change depending on package installation order.  However,
it shelled out twice per file. This is not fast enough if there are many
thousands of files to be restored (in /var/xapi/blobs).

Instead, create a mapping from old id to new id and change the ownership
only when needed, and without shelling out.

With this change, the time to restore the files is approximately the
same as backing them up in the first place.