CA-261633: Fix restoring when using a multipath root
Convert the discovered backup disks to /dev/mapper paths if needed.
Otherwise the paths are of the format, /dev/dm[0-9]+ which the existing
partition tools do not cope with.  Also convert the disk from the
answerfile to a /dev/mapper path if needed so that it is correctly
matched up with the backup disk.