CA-265171: Install bootloader to partition if needed
GRUB requires the first partition to start at least on LBA 63 according
to the documentation.  Experimentally, it is LBA 52 for GRUB 2.02 beta 2
and LBA 53 for GRUB 2.02. This may be an issue during upgrades as the
requirement has increased (especially compared with extlinux).
Automatically install GRUB to the partition if it won't fit in the
post-MBR gap.