CA-263790: Don't remove the wrong volume group during upgrade
If there are many disks connected to the host, checking whether a
partition is part of a disk using a simple substring match may not work
(/dev/sda3 would be considered part of /dev/sda and /dev/sdaa).  An
incorrect match might cause the wrong volume group to be removed during
upgrade. Fix this by actually working out the parent disk of the
partition. The mechanism for doing this is taken from util-linux and is
also used in the host-upgrade-plugin.
if a partition is contained in a disk