CA-266760: Allow upgrading packages from a driver disk
The default behavior of Yum in RHEL7 has group_command=objects which
means that Yum groups are treated as top-level objects which generally
behave more rationally.  However, it doesn't work when installing a
driver disk into the host installer environment if one of the packages
needs to be upgraded because Yum refuses to upgrade a package within a
group if it was installed outside of that group.  Fix this by setting
group_command=compat which will just install/upgrade each package in the
group regardless of origin.