CA-290899: use file system tree rooted directory for rpm query
New version of libnss raises error if /dev/urandom is not available,
given the rpm command uses libnss3.so, we need a extra 'bind /dev'
for executing the chroot command. Instead we can use the file system
tree rooted directory for rpm query by removing the chroot command.

This issue was introduced by:
https://hg.mozilla.org/projects/nss/rev/5ed1226cf418#l3.13

Signed-off-by: Liang Dai <liang.dai1@citrix.com>