CA-310736: iscsi: Don't use gateway if target is on the same subnet
When adding routes to the iSCSI target, if the target IP is on the same
subnet as the initiator IP add a direct route rather than routing via
the gateway. This avoids needlessly routing traffic via the gateway and
also fixes installation if target and initiator are on the same subnet
and the gateway in the iBFT is bogus.