CA-311773: Keep order of updates specified in the answerfile
Since one update may depend on another, maintaining the order of updates
during installation is important. Previously, the order was not
maintained due to moving the main repo to the front and the use of an
unordered set. Fix this by keeping the repos as a list and being careful
to maintain the order of updates.

In the future we could install the updates in the correct order
automatically respecting the ordering of dependent updates.