CA-322479: Prevent iSCSI session(s) getting logged out during shutdown
After installation, prevent iSCSI boot session(s) getting logged out
during shutdown by creating iscsiadm database entries on the root fs.
Set node.startup to "onboot" so that the sesssions are not logged out
during shutdown.

There is a script that runs at boot "iscsi-mark-root-nodes" which is
supposed to mark the sessions corresponding to the root nodes as
"onboot" but this doesn't work because the iscsiadm entries are missing.