CA-323227: Fix interactive installation with multipath
Commit f84ea135c8c2 ("CP-30566: Add support for VirtIO Blk devices")
changed which devices are included by getDiskList() and includes devices
with a major number of 253 directly. This means that /dev/dm-*
device-mapper devices are included in the list which causes other parts
of the installer to barf. Device-mapper devices are already included
specially (via /dev/mapper/*) later in the function so just skip adding
them earlier on.

This fixes interactive installation when multipath is enabled.