CP-30221 / CA-325143: Switch to use chronyd for NTP
Switch to use chronyd and chrony-wait rather than ntpd and ntpdate.
chronyd is recommended to be used by Red Hat over the ntp package and
allows waiting for the daemon to synchronize the time rather than
requiring a separate command to do it.

Also add the "iburst" option to the server directive(s) which is used by
the default server directives and allows time sync to proceed more
quickly during startup.

This also means that ntpdate is no longer started unconditionally which
meant that a manual time configuration would be lost every time the host
is booted.