CP-25099: Set time before installation is started
Set the host's time before starting installation (in the prepare phase).
This ensures that the timestamps of newly created files (e.g. the
initramfs) are all correct and ensures that the time is set correctly
during firstboot. While the system should generally be able to cope with
the time changing, in some cases it doesn't (e.g. dhclient's leases).

For NTP, setting the time is best effort since the network might not be
configured if installing from a disc.

If setting the time manually, the screen to set it is now part of the
pre-installation configuration rather than at the end of installation.