CP-32298: Don't expect firstboot state to exist
Previously, the installer would check the firstboot state directory to
ensure that all firstboot scripts had run before allowing upgrade to
proceed. With the firstboot scripts being removed, this is no longer
possible.

For most of the firstboot scripts, it is irrelevant whether they have
been run before upgrade. However, for a couple of them (network and
storage) it is important because they should only be run after
installation, not upgrade. So check the state files for these particular
services to ensure that they have been run. The old method of checking
firstboot scripts is kept for upgrading from older releases.

This change only checks the storage service. The network service will be
done at a later point.