Unbreak "safe" boot option on x2APIC machines
If a machine supports x2APIC, Xen will try to enable it that will
require IOMMU being enabled later which contradicts the fact we disable
it explicitly here. Furthermore, if x2APIC is preenabled by firmware
there is no way currently to downgrade it meaning Xen will fail to
boot on that machine. In order to fix it, keep IOMMU enabled in
safe mode which should be fine these days.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>