CA-336696: Handle errors when copying ownership
The lines in /etc/passwd and /etc/group may contain an invalid number
of fields or the fields may be invalid (e.g. not a number). Also, the
source username/group may not have a matching destination username/group.
Log a warning rather than failing since it is better to carry on at this
point rather than ending up with a broken system.