CA-336693: Fix UEFI restore when there is no storage partition
If there is no storage partition, the boot partition is 3 instead of 4
but the code didn't detect this properly due to always passing
create_sr_part=True to inspectTargetDisk(). This meant that it would try
to install a BIOS bootloader because the ESP was not correctly detected.
Instead, detect the storage partition and pass the correct value for
create_sr_part.