CA-355255: Don't fail to install over another OS
The logic to fail installation with a DOS utility partition was a bit
overzealous, resulting in an installation failure if the disk has a DOS
partition label even if there is no utility partition present. This may
be the case if installing over another OS.

Instead, block installation only if the disk has a DOS partition label
and a partition would be preserved. In other words:

* Either preserve-first-partition is set to 'true', or
* preserve-first-partition is set to 'if-utility' and there is a utility
  partition present.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>