CA-349118: Create modprobe files during upgrade
Depending on settings, the host-installer will create modprobe blacklist
files for bridge and ipv6 during clean install. These need to be
recreated during upgrade.

Move the creation of the bridge blacklist earlier so that it happens for
upgrades too.
Fix the IPv6 check during upgrade - when IPv6 is not used,
/etc/sysconfig/network may have NETWORKING_IPV6=no so check for that
too, not just its absence.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>