CA-369534: Preserve utility partitions when running interactively
The answerfile allows choosing whether to preserve the first partition.
It defaults to yes if it is a utility partition. For interactive
installs, it is not possible to choose this option but since
'preserve-first-partition' is not specified at all, it has the same
behaviour as "no".

This difference is confusing at a minimum and risks losing the utility
partition if users expect the default to be 'if-utility' as it is
documented for answerfile installs.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>