CA-369534: Fix creating an SR during interactive installs
Commit e5df6f8aaa74 ("CP-32753 - Remove OEM utility partition support")
inadvertently swapped two parameters in inspectTargetDisk(). This means
that create_sr_part is assigned the value of preserve_first_partition.

For answerfile installs, the default 'if-utility' value of
preserve_first_partition is treated as true so the local SR partition is
created.

For interactive installs, preserve_first_partition is always None so the
SR is never created.

Fixes: e5df6f8aaa74 ("CP-32753 - Remove OEM utility partition support")
Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>