CA-369898: Replace PRODUCT_VERSION with PRODUCT_VERSION_TEXT
The product version is a marketing identifier and should not be used for
any version number comparisons. The platform version should be used for
version number comparisons instead - this was fixed by 7613817e230a
("CP-15346: Switch to using PLATFORM_VERSION for upgrade version
checks").

Since PRODUCT_VERSION_TEXT is the preferred display string, use that
throughout the installer. At the same time, remove code which handles
the possibility that the product version is not set since it is always
set. Also drop displaying of the build number since it doesn't work
well with the textual product version and doesn't give any value with
the cloud stream. Ensure that the build number and numeric product
versions of existing installations and backups are logged just in case.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>