init: allow 'answerfile' to work without 'install' boot parameter
Booting the installer with `answerfile=URL` and no `install` to
complement it would just get no effect, with just a cryptic "No
operation specified for answerfile" hidden in the logfile.

The `install` parameter had only one use in the current code: allowing
`answerfile` to be honored, so just ignore it.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>