Fix comment and range for dynamically-assigned device majors
The commit message for f84ea135c8c2fb961541085207a92cd96517dc36
explains that what we want is the dynamic major range, while the code
checks for (and documents) the local/experimental range instead, which
happen to be a significant subset of the former.

This fixes the comment for correctness, and (mostly "just in case" and
to match the intent) extends the check to the full dynamic range, and
adds the second dynamic range.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>
CC: Igor Druzhinin <igor.druzhinin@citrix.com>