RAID: when a MD superblock is found on a disk, offer user to assemble RAID
This adds an option to enable *all* RAID devices found.  This is less
flexible than a more selective mechanism would be, but really simple and
covering all known use-cases.

It filters out from choices presented to the user both "update of" and
"restore to" entries refering to a disk that is part of a md RAID
(though in practice only the "update" case is likely to appear).  This
avoids the case where a user gets those entries before explicitly
assembling RAIDs, and thus avoids breaking a RAID by mistake.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>