WIP mockrun-installer: new script to help in installer development
This allows to test quite some things, the default answers with
`--netinstall` let an upgrade be testable until the `yum install` and
check in the logs the packages that would get installed.

This is likely to require for some time manual edit of the script to
customize anything.

Could show the way to real UI testing, by mocking user input as well.