netutil.scanConfiguration: allow reuse of commandline config
Bonding configuration created from commandline cannot be properly
passed to the UI, which needs it both for netinstall and
host-management, so we have to re-discover the existing configuration
to present it to the user.  This includes identifying the configured
bonding interface, and filtering out their member interfaces from
available choices.

This is not very satisfying, and we may want to improve the overall
installer architecture to make this better in the future, but it would be
way out of scope here.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>