CA-384621: Do not move partition to preserve data
During installation/upgrade we remove all partitions and then create
new partitions as required. This is done also for partitions that
needs to be preserved (like Dell utility partition).
By default partition utilities tries to align start of the
partitions for better performance but this causes partitions with
data to be recreated with different start/end corrupting them.
To avoid the corruption do the alignment from our code when
we create new partitions and avoid alignment done in the utilities
(adding "--set-alignment=1" parameter).
As a minor change create a single command to create and setup
a single partition instead of using multiple commands.

Signed-off-by: Frediano Ziglio <frediano.ziglio@cloud.com>